Update Employee Plugin

+++ Plugins/Nop.Plugin.Widgets.Ecommerce/Areas/Admin/Controller/CompanyController.cs
@@ -5,15 +5,15 @@ using Nop.Plugin.Widgets.Ecommerce.Domain;
 using Nop.Plugin.Widgets.Ecommerce.Services;
 using Nop.Services.Localization;
 using Nop.Services.Media;
+using Nop.Services.Messages;
+using Nop.Services.Security;
+using Nop.Web.Areas.Admin.Controllers;
 using Nop.Web.Areas.Admin.Infrastructure.Mapper.Extensions;
-using Nop.Web.Framework;
-using Nop.Web.Framework.Controllers;
 using Nop.Web.Framework.Mvc.Filters;
 
 namespace Nop.Plugin.Widgets.Ecommerce.Areas.Admin.Controller;
-[AuthorizeAdmin]
-[Area(AreaNames.ADMIN)]
-public class EcommerceCompanyController : BasePluginController
+
+public class CompanyController : BaseAdminController
 {
     #region Fields
 
@@ -21,21 +21,30 @@ public class EcommerceCompanyController : BasePluginController
     private readonly ICompanyModelFactory _companyModelFactory;
     private readonly IPictureService _pictureService;
     private readonly ILocalizedEntityService _localizedEntityService;
+    private readonly INotificationService _notificationService;
+    private readonly IPermissionService _permissionService;
+    private readonly ILocalizationService _localizationService;
 
     #endregion
 
     #region Ctor
-    public EcommerceCompanyController(
+    public CompanyController(
         ICompanyService companyService,
         ICompanyModelFactory companyModelFactory,
         IPictureService pictureService,
-        ILocalizedEntityService localizedEntityService
+        ILocalizedEntityService localizedEntityService,
+        INotificationService notificationService,
+        IPermissionService permissionService,
+        ILocalizationService localizationService
         )
     {
         _companyService = companyService;
         _companyModelFactory = companyModelFactory;
         _pictureService = pictureService;
         _localizedEntityService = localizedEntityService;
+        _notificationService = notificationService;
+        _permissionService = permissionService;
+        _localizationService = localizationService;
     }
 
     #endregion
@@ -76,10 +85,12 @@ public class EcommerceCompanyController : BasePluginController
 
     public async Task<IActionResult> List()
     {
+        if (!await _permissionService.AuthorizeAsync(StandardPermissionProvider.AccessAdminPanel))
+            return AccessDeniedView();
 
         var searchModel = await _companyModelFactory.PrepareCompanySearchModelAsync(new Model.CompanySearchModel());
 
-        return View("List", searchModel);
+        return View(searchModel);
     }
 
 
@@ -87,6 +98,14 @@ public class EcommerceCompanyController : BasePluginController
     [HttpPost]
     public async Task<IActionResult> List(CompanySearchModel searchModel)
     {
+        if (!await _permissionService.AuthorizeAsync(StandardPermissionProvider.AccessAdminPanel))
+            return AccessDeniedView();
+
+        if (!ModelState.IsValid)
+        {
+            return Json(new { Success = false });
+        }
+
         var model = await _companyModelFactory.PrepareCompanyListModelAsync(searchModel);
 
         return Json(model);
@@ -98,9 +117,12 @@ public class EcommerceCompanyController : BasePluginController
 
     public async Task<IActionResult> Create()
     {
+        if (!await _permissionService.AuthorizeAsync(StandardPermissionProvider.AccessAdminPanel))
+            return AccessDeniedView();
+
         var model = await _companyModelFactory.PrepareCompanyModelAsync(new CompanyModel(), null);
 
-        return View("Create", model);
+        return View(model);
     }
 
 
@@ -109,6 +131,9 @@ public class EcommerceCompanyController : BasePluginController
     [HttpPost, ParameterBasedOnFormName("save-continue", "continueEditing")]
     public async Task<IActionResult> Create(CompanyModel model, bool continueEditing)
     {
+        if (!await _permissionService.AuthorizeAsync(StandardPermissionProvider.AccessAdminPanel))
+            return AccessDeniedView();
+
         if (ModelState.IsValid)
         {
             var company = model.ToEntity<Company>();
@@ -117,6 +142,7 @@ public class EcommerceCompanyController : BasePluginController
             await UpdatePictureSeoNamesAsync(company);
 
             await UpdateLocalesAsync(company, model);
+            _notificationService.SuccessNotification(await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.Created"));
 
             return continueEditing ? RedirectToAction("Edit", new { id = company.Id }) : RedirectToAction("List");
         }
@@ -130,24 +156,41 @@ public class EcommerceCompanyController : BasePluginController
 
     public async Task<IActionResult> Edit(int id)
     {
+        if (!await _permissionService.AuthorizeAsync(StandardPermissionProvider.AccessAdminPanel))
+            return AccessDeniedView();
+
         var company = await _companyService.GetCompanyByIdAsync(id);
 
         if (company == null)
+        {
+            var message = await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.NotFound");
+            _notificationService.ErrorNotification(message);
             return RedirectToAction("List");
+        }
 
 
         var model = await _companyModelFactory.PrepareCompanyModelAsync(null, company);
 
-        return View("Edit", model);
+        return View(model);
     }
 
 
     [HttpPost, ParameterBasedOnFormName("save-continue", "continueEditing")]
     public async Task<IActionResult> Edit(CompanyModel model, bool continueEditing)
     {
+
+        if (!await _permissionService.AuthorizeAsync(StandardPermissionProvider.AccessAdminPanel))
+            return AccessDeniedView();
+
+
         var company = await _companyService.GetCompanyByIdAsync(model.Id);
         if (company == null)
+        {
+
+            var message = await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.EditFailed");
+            _notificationService.ErrorNotification(message);
             return RedirectToAction("List");
+        }
 
 
         if (ModelState.IsValid)
@@ -159,12 +202,12 @@ public class EcommerceCompanyController : BasePluginController
 
             await UpdatePictureSeoNamesAsync(company);
             await UpdateLocalesAsync(company, model);
-
+            _notificationService.SuccessNotification(await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.UpdateSuccessfully"));
             return continueEditing ? RedirectToAction("Edit", new { id = company.Id }) : RedirectToAction("List");
         }
 
         model = await _companyModelFactory.PrepareCompanyModelAsync(model, company);
-        return View("Edit", model);
+        return View(model);
     }
 
 
@@ -173,11 +216,18 @@ public class EcommerceCompanyController : BasePluginController
     [HttpPost]
     public async Task<IActionResult> Delete(CompanyModel model)
     {
+        if (!await _permissionService.AuthorizeAsync(StandardPermissionProvider.AccessAdminPanel))
+            return AccessDeniedView();
+
         var company = await _companyService.GetCompanyByIdAsync(model.Id);
         if (company == null)
+        {
+            _notificationService.ErrorNotification(await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.NotFound"));
             return RedirectToAction("List");
+        }
 
         await _companyService.DeleteCompanyAsync(company);
+        _notificationService.SuccessNotification(await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.Deleted"));
         return RedirectToAction("List");
     }
 
@@ -189,19 +239,22 @@ public class EcommerceCompanyController : BasePluginController
 
 
         if (selectedId == null || !selectedId.Any())
-            return NoContent();
+        {
+            _notificationService.ErrorNotification(await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.NotFound"));
+            return RedirectToAction("List");
+        }
         try
         {
-
             foreach (var id in selectedId)
             {
                 var company = await _companyService.GetCompanyByIdAsync(id);
                 if (company != null)
                 {
                     await _companyService.DeleteCompanyAsync(company);
-
                 }
             }
+
+            _notificationService.SuccessNotification(await _localizationService.GetResourceAsync("Nop.Plugin.Widget.Ecommerce.Company.Deleted"));
         }
         catch (Exception)
         {
